+++ protected/components/DatasetPageSettings.php
@@ -33,7 +33,7 @@ class DatasetPageSettings extends yii\base\BaseObject
         parent::__construct();
         $this->_dao = $dao;
         $this->_model = $model;
-        if ($this->_model && !$this->_model->isNewRecord && in_array($this->_model->upload_status, ["ImportFromEM","UserUploadingData","DataAvailableForReview", "DataPending", "Curation", "AuthorReview"])) {
+        if ($this->_model && !$this->_model->isNewRecord && in_array($this->_model->upload_status, ["ImportFromEM", "UserUploadingData", "UserProvidedData", "DataPreparation", "Curation", "AuthorReview"])) {
             $this->_pageType = "draft";
         } elseif (
             $this->_model && !$this->_model->isNewRecord && "Published" !== $this->_model->upload_status